+++ changelog.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * New `Slowest` Queries panel for cborm, acf, and qb/quick
 * New visualizer total db time as well as request time including percentage of the request time
 * Ability to export a profiler in json
+* Ability to sort the visualizer's profilers
 
 ### Fixed